refactor: update dao with interface
