+++ src/main/java/ru/clevertec/dao/tag/TagDAO.java
@@ -1,10 +1,11 @@
-package ru.clevertec.dao;
+package ru.clevertec.dao.tag;
 
+import ru.clevertec.dao.DAOInterface;
 import ru.clevertec.entity.Tag;
 
 import java.util.List;
 
-public interface TagDAO {
+public interface TagDAO extends DAOInterface<Tag> {
 
     void create(Tag tag);
